refactor: split entries/models.py into a package by entry category
