+++ camp/apps/entries/models/base.py
@@ -12,9 +12,9 @@ from django_smalluuid.models import SmallUUIDField, uuid_default
 from camp.apps.monitors.models import Monitor
 from camp.utils import classproperty
 
-from . import stages
-from .levels import LevelSet, AQLevel
-from .managers import EntryQuerySet
+from ..levels import LevelSet, AQLevel
+from ..managers import EntryQuerySet
+from .. import stages
 
 
 class BaseEntry(models.Model):
@@ -299,247 +299,3 @@ class BaseEntry(models.Model):
             readings[key] = entry.declared_data()
 
         return readings
-
-
-# Particulate Matter
-
-class PM25(BaseEntry):
-    label = _('PM2.5')
-    epa_aqs_code = 88101
-    units = 'µg/m³'
-    summarize = True
-
-    Levels = LevelSet(
-        AQLevel.GOOD(0.0),
-        AQLevel.MODERATE(9.1),
-        AQLevel.UNHEALTHY_SENSITIVE(35.5),
-        AQLevel.UNHEALTHY(55.5),
-        AQLevel.VERY_UNHEALTHY(150.5),
-        AQLevel.HAZARDOUS(250.5),
-    )
-
-    value = models.DecimalField(
-        max_digits=7, decimal_places=2,
-        help_text=_('PM2.5 (µg/m³)'),
-    )
-
-
-
-class Particulates(BaseEntry):
-    label = _('Particulates')
-    particles_03um = models.DecimalField(max_digits=8, decimal_places=2, null=True)
-    particles_05um = models.DecimalField(max_digits=8, decimal_places=2, null=True)
-    particles_10um = models.DecimalField(max_digits=8, decimal_places=2, null=True)
-    particles_25um = models.DecimalField(max_digits=8, decimal_places=2, null=True)
-    particles_50um = models.DecimalField(max_digits=8, decimal_places=2, null=True)
-    particles_100um = models.DecimalField(max_digits=8, decimal_places=2, null=True)
-
-
-class PM10(BaseEntry):
-    label = _('PM1.0')
-    units = 'µg/m³'
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text='PM1.0 (µg/m³)'
-    )
-
-
-class PM100(BaseEntry):
-    label = _('PM10.0')
-    epa_aqs_code = 81102
-    units = 'µg/m³'
-
-    Levels = LevelSet(
-        AQLevel.GOOD(0),
-        AQLevel.MODERATE(55),
-        AQLevel.UNHEALTHY_SENSITIVE(155),
-        AQLevel.UNHEALTHY(255),
-        AQLevel.VERY_UNHEALTHY(355),
-        AQLevel.HAZARDOUS(425),
-        AQLevel.VERY_HAZARDOUS(605),
-    )
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text='PM10.0 (µg/m³)'
-    )
-
-
-# Meteorological
-
-class Temperature(BaseEntry):
-    label = _('Temperature')
-    epa_aqs_code = 62101
-    units = '°F'
-
-    value = models.DecimalField(
-        max_digits=4, decimal_places=1,
-        help_text=_('Temperature (°F)')
-    )
-
-    @property
-    def fahrenheit(self):
-        return self.value
-
-    @fahrenheit.setter
-    def fahrenheit(self, value):
-        self.value = value
-
-    @property
-    def celsius(self):
-        value = (Decimal(self.fahrenheit) - 32) * (Decimal(5) / Decimal(9))
-        return value.quantize(Decimal('0.1'), rounding=ROUND_HALF_UP)
-
-    @celsius.setter
-    def celsius(self, value):
-        value = (Decimal(value) * (Decimal(9) / Decimal(5))) + 32
-        self.fahrenheit = value.quantize(Decimal('0.1'), rounding=ROUND_HALF_UP)
-
-
-    def serialized_data(self):
-        return {
-            'temperature_f': self.fahrenheit,
-            'temperature_c': self.celsius,
-        }
-
-
-class Humidity(BaseEntry):
-    label = _('Humidity')
-    epa_aqs_code = 62201
-    units = '%'
-
-    value = models.DecimalField(
-        max_digits=4, decimal_places=1,
-        help_text=_('Relative humidity (%)')
-    )
-
-
-class Pressure(BaseEntry):
-    label = _('Atmospheric Pressure')
-    units = 'mmHg'
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text=_('Atmospheric pressure (mmHg)'),
-    )
-
-    @property
-    def mmhg(self):
-        return self.value
-
-    @mmhg.setter
-    def mmhg(self, value):
-        self.value = Decimal(value).quantize(Decimal('0.01'), rounding=ROUND_HALF_UP)
-
-    @property
-    def hpa(self):
-        return (self.mmhg * Decimal('1.33322')).quantize(Decimal('0.1'), rounding=ROUND_HALF_UP)
-
-    @hpa.setter
-    def hpa(self, value):
-        self.mmhg = Decimal(value) / Decimal('1.33322')
-
-    def serialized_data(self):
-        return {
-            'pressure_mmhg': self.mmhg,
-            'pressure_hpa': self.hpa,
-        }
-
-
-# Gases
-
-class CO(BaseEntry):
-    label = _('Carbon Monoxide')
-    epa_aqs_code = 42101
-    units = 'ppm'
-    summarize = True
-
-    Levels = LevelSet(
-        AQLevel.GOOD(0.0),
-        AQLevel.MODERATE(4.5),
-        AQLevel.UNHEALTHY_SENSITIVE(9.5),
-        AQLevel.UNHEALTHY(12.5),
-        AQLevel.VERY_UNHEALTHY(15.5),
-        AQLevel.HAZARDOUS(30.5),
-        AQLevel.VERY_HAZARDOUS(50.4),
-    )
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text='Carbon monoxide (ppm)',
-    )
-
-
-class CO2(BaseEntry):
-    label = _('Carbon Dioxide')
-    epa_aqs_code = 42102
-    units = 'ppm'
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text='Carbon dioxide (ppm)',
-    )
-
-
-class NO2(BaseEntry):
-    label = _('Nitrogen Dioxide')
-    epa_aqs_code = 42602
-    units = 'ppb'
-    summarize = True
-
-    Levels = LevelSet(
-        AQLevel.GOOD(0.0),
-        AQLevel.MODERATE(54.0),
-        AQLevel.UNHEALTHY_SENSITIVE(101.0),
-        AQLevel.UNHEALTHY(361.0),
-        AQLevel.VERY_UNHEALTHY(650.0),
-        AQLevel.HAZARDOUS(1250.0),
-        AQLevel.VERY_HAZARDOUS(2050.0),
-    )
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text='Nitrogen dioxide (ppb)',
-    )
-
-
-class O3(BaseEntry):
-    label = _('Ozone')
-    epa_aqs_code = 44201
-    units = 'ppb'
-    summarize = True
-
-    Levels = LevelSet(
-        AQLevel.GOOD(0.0),
-        AQLevel.UNHEALTHY_SENSITIVE(125),
-        AQLevel.UNHEALTHY(165),
-        AQLevel.VERY_UNHEALTHY(205),
-        AQLevel.HAZARDOUS(405),
-        AQLevel.VERY_HAZARDOUS(605),
-    )
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text='Ozone (ppb)'
-    )
-
-
-class SO2(BaseEntry):
-    label = _('Sulfur Dioxide')
-    epa_aqs_code = 42401
-    units = 'ppb'
-    summarize = True
-
-    Levels = LevelSet(
-        AQLevel.GOOD(0),
-        AQLevel.MODERATE(36),
-        AQLevel.UNHEALTHY_SENSITIVE(76),
-        AQLevel.UNHEALTHY(186),
-        AQLevel.VERY_UNHEALTHY(305),
-    )
-
-    value = models.DecimalField(
-        max_digits=6, decimal_places=2,
-        help_text=_('Sulfur dioxide (ppb)'),
-    )
